Use our own (private) fork
This action isn't "Verified by GitHub", so I didn't feel too comfortable
entrusting it with a firebase token. In writing this commit I've
noticed that the original author has a github email address (and the
initial commits predate GitHub Actions public launch).

Possibly not necessary to maintain our own fork of this plumbing in the
long term, but thought it would be nice to have full control over the
action as we dabble for the first time.